+++ src/tikzDevice.c
@@ -306,6 +306,10 @@ static Rboolean TikZ_Setup(
   tikzInfo->onefile = onefile;
   tikzInfo->pageNum = 1;
 
+  /* initialize strings, just to be on the safe side */
+  strcpy(tikzInfo->drawColor, "drawColor");
+  strcpy(tikzInfo->fillColor, "fillColor");
+
   /* Incorporate tikzInfo into deviceInfo. */
   deviceInfo->deviceSpecific = (void *) tikzInfo;
 
@@ -508,13 +512,6 @@ static Rboolean TikZ_Setup(
 static void TikZ_WriteColorDefinition( tikzDevDesc *tikzInfo, void (*printOut)(tikzDevDesc *tikzInfo, const char *format, ...), int color, const char* colorname, const char* colorstr )
 {
 
-  /* strip # from both strings as they are not necessary nor allowed */
-  if( colorname[0] == '#' )
-    colorname= colorname +1;
-
-  if( colorstr[0] == '#')
-    colorstr = colorstr+1;
-
   /* define the color as an alias */
   if( color == -1 )
     printOutput(tikzInfo,
@@ -546,6 +543,10 @@ static void TikZ_WriteColorDefinitions( tikzDevDesc *tikzInfo )
   for( i = 0; i < tikzInfo->ncolors; ++i)
   {
     const char* colorstr = col2name(tikzInfo->colors[i]);
+
+    if(colorstr[0] == '#')
+      colorstr = colorstr+1;
+
     TikZ_WriteColorDefinition(tikzInfo, printColorOutput, tikzInfo->colors[i], colorstr, colorstr);
   }
 }
@@ -1153,7 +1154,7 @@ static void TikZ_Text( double x, double y, const char *str,
   TikZ_DefineColors(plotParams, deviceInfo, DRAWOP_DRAW);
 
   /* Start a node for the text, open an options bracket. */
-  printOutput(tikzInfo,"\n\\node[text=drawColor");
+  printOutput(tikzInfo,"\n\\node[text=%s", tikzInfo->drawColor);
   /* FIXME: Should bail out of this function early if text is fully transparent */
   if( !R_OPAQUE(plotParams->col) )
     printOutput(tikzInfo, ",text opacity=%4.2f", R_ALPHA(plotParams->col)/255.0);
@@ -1787,14 +1788,35 @@ static Rboolean TikZ_CheckAndAddColor(tikzDevDesc *tikzInfo, int color)
   return colorfound;
 }
 
-static void TikZ_DefineDrawColor(tikzDevDesc *tikzInfo, int color, const char* colortype)
+static void TikZ_DefineDrawColor(tikzDevDesc *tikzInfo, int color, TikZ_DrawOps ops)
 {
   const char *colorstr = col2name(color);
+  const char* colors[] = {"", "drawColor", "fillColor"};
+  char* dest;
+
+  if( colorstr[0] == '#' )
+    colorstr = colorstr+1;
+
+  if( ops == DRAWOP_DRAW)
+  {
+    dest = tikzInfo->drawColor;
+  }
+  else if( ops == DRAWOP_FILL )
+  {
+    dest = tikzInfo->fillColor;
+  }
 
   if( TikZ_CheckAndAddColor(tikzInfo, color) )
-    TikZ_WriteColorDefinition(tikzInfo, printOutput, -1, colortype, colorstr);
+  {
+    strlcpy(dest, colorstr, strlen(colorstr));
+  }
   else
-    TikZ_WriteColorDefinition(tikzInfo, printOutput, color, colortype, colorstr);
+  {
+    strlcpy(dest, colors[ops], strlen(colorstr));
+
+    TikZ_WriteColorDefinition(tikzInfo, printOutput, color, colors[ops], colorstr);
+  }
+
 
 }
 
@@ -1805,10 +1827,9 @@ static void TikZ_DefineColors(pGEcontext plotParams, pDevDesc deviceInfo, TikZ_D
 
   if ( ops & DRAWOP_DRAW ) {
     color = plotParams->col;
-
     if ( color != tikzInfo->oldDrawColor ) {
       tikzInfo->oldDrawColor = color;
-      TikZ_DefineDrawColor(tikzInfo, color, "drawColor");
+      TikZ_DefineDrawColor(tikzInfo, color, DRAWOP_DRAW);
     }
   }
 
@@ -1816,7 +1837,7 @@ static void TikZ_DefineColors(pGEcontext plotParams, pDevDesc deviceInfo, TikZ_D
     color = plotParams->fill;
     if( color != tikzInfo->oldFillColor ) {
       tikzInfo->oldFillColor = color;
-      TikZ_DefineDrawColor(tikzInfo, color, "fillColor");
+      TikZ_DefineDrawColor(tikzInfo, color, DRAWOP_FILL);
     }
   }
 
@@ -1838,7 +1859,7 @@ static void TikZ_WriteDrawOptions(const pGEcontext plotParams, pDevDesc deviceIn
   tikzDevDesc *tikzInfo = (tikzDevDesc *) deviceInfo->deviceSpecific;
 
   if ( ops & DRAWOP_DRAW ) {
-    printOutput(tikzInfo, "draw=drawColor");
+    printOutput(tikzInfo, "draw=%s", tikzInfo->drawColor);
     if( !R_OPAQUE(plotParams->col) )
       printOutput(tikzInfo, ",draw opacity=%4.2f", R_ALPHA(plotParams->col)/255.0);
 
@@ -1850,7 +1871,7 @@ static void TikZ_WriteDrawOptions(const pGEcontext plotParams, pDevDesc deviceIn
     if ( ops & DRAWOP_DRAW )
       printOutput(tikzInfo, ",");
 
-    printOutput(tikzInfo, "fill=fillColor");
+    printOutput(tikzInfo, "fill=%s", tikzInfo->fillColor);
     if( !R_OPAQUE(plotParams->fill) )
       printOutput(tikzInfo, ",fill opacity=%4.2f", R_ALPHA(plotParams->fill)/255.0);
   }
@@ -2259,12 +2280,12 @@ static void TikZ_CheckState(pDevDesc deviceInfo)
      */
     int color = deviceInfo->startfill;
     tikzInfo->oldFillColor = color;
-    TikZ_DefineDrawColor(tikzInfo, color, "fillColor");
+    TikZ_DefineDrawColor(tikzInfo, color, DRAWOP_FILL);
 
     printOutput(tikzInfo, "\\path[use as bounding box");
 
     /* TODO: Consider only filling when the color is not transparent. */
-    printOutput(tikzInfo, ",fill=fillColor");
+    printOutput(tikzInfo, ",fill=%s", tikzInfo->fillColor);
     if( !R_OPAQUE(color) )
       printOutput(tikzInfo, ",fill opacity=%4.2f", R_ALPHA(color)/255.0);